added user FK in todos table, updated schema.sql file with users table

+++ controllers/todos.js
@@ -2,7 +2,7 @@ const pool = require('../db/index.js')
 
 const getAllTodo = async (req, res,next) => {
     try {
-        const result = await pool.query('SELECT * FROM todos')
+        const result = await pool.query('SELECT * FROM todos WHERE user_id = $1', [req.user.id])
         res.json(result.rows)
     } catch (error) {
         next(error)
@@ -12,7 +12,7 @@ const getAllTodo = async (req, res,next) => {
 const getTodoById = async (req, res, next) => {
     const id = req.params.id
     try {
-        const result = await pool.query('SELECT * FROM todos WHERE id = $1', [id])
+        const result = await pool.query('SELECT * FROM todos WHERE id = $1 AND user_id = $2', [id, req.user.id])
         if(!result.rows[0]){
             const err = new Error("Requested todo not found")
             err.status = 404
@@ -38,7 +38,7 @@ const postTodo = async (req, res, next) => {
     }
 
     try {
-        const result = await pool.query('INSERT INTO todos (title, description, completed) VALUES ($1, $2, $3) RETURNING *', [req.body.title, req.body.description, false])
+        const result = await pool.query('INSERT INTO todos (title, description, completed, user_id) VALUES ($1, $2, $3, $4) RETURNING *', [req.body.title, req.body.description, false, req.user.id])
         return res.status(201).json(result.rows[0])
     } catch (error) {
         next(error)
@@ -48,7 +48,7 @@ const postTodo = async (req, res, next) => {
 const putTodo = async (req, res, next) => {
     const id = req.params.id
     try {
-        const result = await pool.query('UPDATE todos SET title = $1, description = $2, completed = $3 WHERE id = $4 RETURNING *', [req.body.title, req.body.description, req.body.completed, id])
+        const result = await pool.query('UPDATE todos SET title = $1, description = $2, completed = $3 WHERE id = $4 AND user_id = $5 RETURNING *', [req.body.title, req.body.description, req.body.completed, id, req.user.id])
         if(!result.rows[0]){
             const err = new Error("Requested todo not found")
             err.status=404
@@ -63,7 +63,7 @@ const putTodo = async (req, res, next) => {
 const deleteTodo = async (req, res, next) => {
     const id = req.params.id
     try {
-        const result = await pool.query('DELETE FROM todos WHERE id = $1 RETURNING *', [id])
+        const result = await pool.query('DELETE FROM todos WHERE id = $1 AND user_id = $2 RETURNING *', [id, req.user.id])
         if(!result.rows[0]){
             const err = new Error("Requested todo not found")
             err.status=404
